feat(Money.java): 금액 검증을 한다
