+++ src/main/java/lotto/model/LottoQuantityChecker.java
@@ -3,8 +3,7 @@ package lotto.model;
 public class LottoQuantityChecker {
     private static final int LOTTO_PRICE = 1000;
 
-    public static int calculate(String source) {
-        int price = Integer.parseInt(source);
-        return price / LOTTO_PRICE;
+    public static int calculate(Money money) {
+        return money.getMoney() / LOTTO_PRICE;
     }
 }
